Move sheet specific helper functions to separate module
